CSS - Rework App And LocationGrid layouts
A major refactor of the existing CSS, continuing the work of the
previous refactor and elminating the need of any custom properties
being set in LocationGrid.tsx.  As well as this, most instances
of custom properties have been removed, leaving only ones for the
sake of clarity or for necessity.

LocationGrid.css now uses flexbox to dynamically determine the widths,
heights, and positions of elements within given the fixed size of the
whole grid.  The only other relevant dimension, that of the dots and
the width of the lines connecting them representing moves, is now
just based off of ems instead of being placed in a custom property.

Padding as well, since so many calculations have been removed from
LocationGrid.css and instead are now done implicitly by the browser's
layout engine, no longer needs to be a property and can instead be
set for all children divs of the App in App.css