LocationGrid - Yet Another Change To Layout Strat
The previous commit had changed the way that the selection areas
around the dots are calculated to be hard-coded into the CSS in
em's.  This was done because previously, had they had been coded as
percentages of their parent container's width and height instead,
it would lead to issues where the areas to select the dots at the
right and bottom sides of the LocationGrid would be compressed
vertically, horizontally, or both.

However, the previous percentage based radius calculation from
commit b35b7a4 has been restored; for the locations corresponding
to the last row and column of dots, they instead are created in
the locations adjacent to them in the LocationGrid's interior and
translated to their proper locations.  This ensures that they
properly scale with the size of the locations in the LocationGrid